Update e2e testing to use sandbox (#71)
* Update e2e testing to use sandbox

This PR changes the e2e test to target sandbox machines. It's not completely clear yet how this will affect browserstack tests' flakiness (RTT between the VM and the M-Lab server might increase a lot) but this should prevent alerts firing in staging due to lower-than-usual test volume whenever the daily e2e tests fail for a day or two in a row.